Add tests for Windows
