+++ .github/workflows/ci.yml
@@ -18,7 +18,8 @@ jobs:
       fail-fast: false
       matrix:
         gradle-version: ["8.4", "8.5", "8.6", "8.7", "8.8", "8.9", "8.10", "8.11", "8.12"]
-    runs-on: ubuntu-latest
+        os: [ubuntu-latest, windows-latest]
+    runs-on: ${{ matrix.os }}
     steps:
       - uses: actions/checkout@v4
