+++ gradle/libs.versions.toml
@@ -3,10 +3,12 @@ axion-release = "1.18.16"
 plugin-publish = "1.3.1"
 kotest = "5.9.1"
 junit-jupiter = "5.11.4"
+test-logger = "4.0.0"
 
 [plugins]
 plugin-publish = { id = "com.gradle.plugin-publish", version.ref = "plugin-publish" }
 axion-release = { id = "pl.allegro.tech.build.axion-release", version.ref = "axion-release" }
+test-logger = { id = "com.adarshr.test-logger", version.ref = "test-logger" }
 
 [libraries]
 kotest-assertions = { group = "io.kotest", name = "kotest-assertions-core-jvm", version.ref = "kotest" }
